Copy file as a fallback if 'os.link' doesn't work.
In rare cases (e.g.: working in a virtual machine on Linux with the
working directory being a mounted as a folder shared with the Windows
host) the check for 'is_windows' doesn't suffice.
In those situations 'os.link' will throw an exception which can be
handled by falling back to simply copying the data.